Extract ParameterForm into a standalone file.
