+++ src/main-layout/SessionWindow/SessionTab/SessionManagement/SessionParamsForm.tsx
@@ -2,19 +2,16 @@ import {
     PlusOutlined,
     DeleteOutlined
 } from '@ant-design/icons';
-import { Form, Button, Popover, Select, Input } from 'antd';
-import DatePicker from 'antd/lib/date-picker';
-import React, { useRef, useState } from 'react';
+import { Button, Popover } from 'antd';
+import React from 'react';
 import { FixSession } from 'src/services/fix/FixSession';
 import { LM } from 'src/translations/language-manager';
 import "./ParamsForm.scss";
-
-const { Option } = Select;
+import { ParameterForm } from "../../../../common/ParameterForm";
 
 const getIntlMessage = (msg: string, options?: any) => {
     return LM.getMessage(`session_params.${msg}`, options);
 }
-
 
 interface SessionParamsFormProps {
     session: FixSession;
@@ -24,105 +21,6 @@ interface SessionParamsFormState {
     initialized: boolean;
     addFormVisible: boolean;
 }
-
-
-export const ParameterForm = ({ onChange, session, togglePopover }: {
-    onChange: (key: string, value: any) => void,
-    togglePopover: (state: boolean) => void,
-    session?: FixSession
-}) => {
-    const formRef: any = useRef(null);
-    const [variableType, setVariableType] = useState("string")
-
-    const checkFormHasErrors = (): boolean => {
-        const fields = formRef.current?.getFieldsError() ?? [];
-
-        for (let i = 0; i < fields.length; i++) {
-            const field = fields[i];
-            if (field.errors.length > 0) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    let input: any;
-
-    switch (variableType) {
-        case "boolean":
-            input = <Select>
-                <Option key={"1"} value={""}>{""}</Option>
-                <Option key={"Y"} value="Y">Y</Option>
-                <Option key={"N"} value="N">N</Option>
-            </Select>
-            break;
-        case "utctimestamp":
-            input = <DatePicker showTime format="YYYY-MM-DD hh:mm:ss:ms" />
-            break;
-        case "monthyear":
-            input = <DatePicker picker="month" />
-            break;
-        case "utcdateonly":
-            input = <DatePicker format="YYYY-MM-DD" />
-            break;
-        case "utctimeonly":
-            input = <DatePicker picker="time" />
-            break;
-        default:
-            input = <Input />;
-    }
-
-    return (<div className="params-form-container">
-        <div className="header">
-            <div className="close" onClick={() => togglePopover(false)}>✕</div>
-        </div>
-        <div className="data-type-selector">
-            <Form.Item label={getIntlMessage("select_type")} labelCol={{ span: 24 }}>
-                <Select value={variableType} onChange={(value) => {
-                    setVariableType(value);
-                    formRef.current?.setFieldsValue({ "value": undefined })
-                }}>
-                    <Option key={"string"} value={"string"}>String</Option>
-                    <Option key={"boolean"} value={"boolean"}>Boolean</Option>
-                    <Option key={"utctimestamp"} value={"utctimestamp"}>Timestamp</Option>
-                    <Option key={"monthyear"} value={"monthyear"}>Month Year</Option>
-                    <Option key={"utcdateonly"} value={"utcdateonly"}>UTC Date Only</Option>
-                    <Option key={"utctimeonly"} value={"utctimeonly"}>UTC Time Only</Option>
-                </Select>
-            </Form.Item>
-        </div>
-        <Form ref={formRef} initialValues={{ state: session?.profile.autoLoginEnabled, loginMsg: session?.profile.autoLoginMsg }} layout="vertical" className="save-as-form"
-            onFinish={(values) => { onChange(values.key, values.value) }}>
-            <div className="form-item-container">
-                <Form.Item name="key" rules={[{
-                    required: true,
-                }]} label={getIntlMessage("key")}>
-                    <Input />
-                </Form.Item>
-
-                <Form.Item name="value" rules={[{
-                    required: true,
-                }]} label={getIntlMessage("value")}>
-                    {input}
-                </Form.Item>
-            </div>
-            <div style={{ textAlign: "center" }}>
-                <Button className="button-v2" type="primary" style={{ marginLeft: "auto" }}
-                    htmlType="submit" onClick={() => {
-                        setTimeout(() => {
-                            if (!checkFormHasErrors()) {
-                                togglePopover(false)
-                            }
-                        }, 10)
-                    }}>
-                    {getIntlMessage("save").toUpperCase()}
-                </Button>
-            </div>
-        </Form>
-    </div>);
-}
-
 
 export class SessionParamsForm extends React.Component<SessionParamsFormProps, SessionParamsFormState> {
     fieldIterationIndex = 0;
@@ -170,7 +68,7 @@ export class SessionParamsForm extends React.Component<SessionParamsFormProps, S
             {initialized && <React.Fragment>
                 <div className="params-header">
                     <Popover
-                        content={addFormVisible ? <ParameterForm togglePopover={this.togglePopover} session={session} onChange={(key, value) => {
+                        content={addFormVisible ? <ParameterForm messageFunc={getIntlMessage} togglePopover={this.togglePopover} session={session} onChange={(key, value) => {
                             session.setSessionParameter(key, value);
                             this.loadAll();
                         }} /> : null}
